requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-admin-honeypot==1.1.0
django-autoslug==1.9.9
django-crontab==0.7.1
django-jalali==6.2.1
django-jazzmin==3.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
